pyproject.toml
[project]
name = "NovelPanel"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "aiofiles>=25.1.0",
  "civitai-py>=0.1.10",
  "docx2txt>=0.8",
  "fastapi>=0.115.0",
  "fastapi-mcp>=0.4.0",
  "flet-toast>=0.6.0",
  "flet-video>=0.1.1",
  "flet[all]==0.28.3",
  "httpx>=0.28.1",
  "langchain>=1.0.2",
  "langchain-community>=0.3.0",
  "langchain-ollama>=1.0.0",
  "langchain-openai>=1.0.1",
  "langgraph>=1.0.1",
  "loguru>=0.7.3",
  "pillow>=11.3.0",
  "pydantic-settings>=2.11.0",
  "pylint>=4.0.2",
  "pypdf>=5.1.0",
  "pytest>=8.4.2",
  "setuptools>=80.9.0",
  "sqlmodel>=0.0.27",
  "uvicorn>=0.34.0",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "NovelPanel"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"

[dependency-groups]
dev = [
    "flet[all]==0.28.3",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}
